# src/docs/openapi.yaml
openapi: '3.1.0'
info:
  title: BorealisMark Protocol API
  version: '1.2.0'
  description: |
    Blockchain-anchored AI Trust Infrastructure on Hedera Hashgraph.

    BorealisMark provides cryptographic audit certificates, trust scoring,
    and decentralized trust deposit verification for AI agents. All certificates are immutably
    anchored to Hedera Consensus Service (HCS).

    ## Authentication

    Two authentication methods are supported:

    - **JWT Bearer Token** — For dashboard users. Obtain via `POST /v1/auth/login`.
    - **API Key** — For programmatic access. Pass via `X-Api-Key` header or `Authorization: Bearer <key>`.

    ## Rate Limits

    Rate limits vary by API tier. Response headers include:
    - `X-RateLimit-Limit` — Monthly quota
    - `X-RateLimit-Remaining` — Requests remaining
    - `X-RateLimit-Reset` — Unix timestamp when quota resets

    ## API Tiers

    | Tier | Requests/Month | Price |
    |------|---------------|-------|
    | Free | 100 | $0 |
    | Starter | 10,000 | $49/mo |
    | Business | 100,000 | $199/mo |
    | Enterprise | 1,000,000 | $499/mo |
  contact:
    name: BorealisMark Support
    url: https://borealisprotocol.ai
    email: support@borealisprotocol.ai
  license:
    name: Proprietary
    url: https://borealisprotocol.ai/terms
servers:
  - url: https://borealismark-api.onrender.com
    description: Production
  - url: http://localhost:3001
    description: Local Development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /v1/auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key with appropriate scopes
...
paths:
...
  /v1/webhooks:
    post:
      tags: [Webhooks]
      summary: Register a webhook endpoint
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                    enum: [audit.completed, audit.anchored, score.degraded, score.improved, trust_deposit.allocated, penalty.executed, agent.registered, key.revoked, webhook.test]
      responses:
        '201':
          description: Webhook registered (secret returned once)
    get:
      tags: [Webhooks]
      summary: List registered webhooks
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Webhook list
  /v1/webhooks/{id}/test:
    post:
      tags: [Webhooks]
      summary: Send test event to webhook
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Test event dispatched
  /v1/webhooks/{id}:
    delete:
      tags: [Webhooks]
      summary: Delete a webhook
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook deleted

  # ── Trust Deposit ─────────────────────────────────────────────────
...